+++ src/server/routers/player/router.ts
@@ -1,9 +1,11 @@
 import { router } from "../../trpc"
-import { getIsInGame } from "./endpoints"
+import { getAllData, getIcon, getIsInGame } from "./endpoints"
 import { liveGameRouter } from "./liveGame/router"
 
 const playerRouter = router({
 	getIsInGame,
+	getIcon,
+	getAllData,
 	liveGame: liveGameRouter,
 })